Refactored PIX component
